feat: enhance posts screen with real characters and improve profile navigation
- Update PostsReelsScreen to use real character data from characterTypes
- Enhance UserProfileScreen with detailed character information
- Fix shadow props deprecation warnings by adding Platform-specific styling
- Fix colors import error in fakeData.ts
- Translate all Hebrew comments to professional English
- Improve navigation between posts and user profiles
- Add character-specific post content and descriptions
- Display verification badges, roles, interests, and activity stats in profiles
