+++ components/PostsReelsScreen.tsx
@@ -9,6 +9,7 @@ import {
   TouchableOpacity,
   Alert,
   Share,
+  Platform,
 } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import Animated, { useAnimatedStyle, withTiming } from 'react-native-reanimated';
@@ -18,6 +19,7 @@ import CommentsModal from './CommentsModal';
 import { isBookmarked, addBookmark, removeBookmark } from '../utils/bookmarksService';
 import colors from '../globals/colors';
 import { FontSizes } from '../globals/constants';
+import { characterTypes, CharacterType } from '../globals/characterTypes';
 
 const { width } = Dimensions.get('window');
 
@@ -46,35 +48,64 @@ type Item = {
 };
 
 /**
- * יוצר נתונים מדומים לפוסטים ורילס
- * @returns מערך של פריטים עם נתונים אקראיים
+ * יוצר נתונים מדומים לפוסטים ורילס עם הדמיות האמיתיות שלנו
+ * @returns מערך של פריטים עם נתונים מהדמיות
  */
 const generateFakeData = (): Item[] => {
   const data: Item[] = [];
-  const userNames = ['אנה כהן', 'דני לוי', 'שרה אברהם', 'משה דוד', 'רחל גולדברג', 'יוסי שפירא'];
+  
+  // פוסטים אמיתיים מהדמיות שלנו
+  const postTopics = [
+    'תרומה לקהילה', 'התנדבות השבוע', 'שיתוף ידע', 'בקשת עזרה', 'הודיה לקהילה',
+    'אירוע קהילתי', 'טיפ מועיל', 'חוויה אישית', 'פרויקט חדש', 'הישג אישי',
+    'בקשת ייעוץ', 'שיתוף חוויה', 'הזמנה לפעילות', 'עדכון פרויקט', 'תודות לקהילה'
+  ];
   
   for (let i = 1; i <= NUM_ITEMS; i++) {
     const type = Math.random() < 0.5 ? 'post' : 'reel';
-    const randomUser = userNames[Math.floor(Math.random() * userNames.length)];
-    const randomLikes = Math.floor(Math.random() * 500) + 10;
-    const randomComments = Math.floor(Math.random() * 50) + 1;
+    const randomCharacter = characterTypes[Math.floor(Math.random() * characterTypes.length)];
+    const randomTopic = postTopics[Math.floor(Math.random() * postTopics.length)];
+    const randomLikes = Math.floor(Math.random() * (randomCharacter.followersCount * 0.1)) + 5;
+    const randomComments = Math.floor(Math.random() * 30) + 1;
+    
+    // יצירת תיאור ייחודי לכל דמות
+    const getCharacterSpecificDescription = (character: CharacterType, topic: string) => {
+      const descriptions = {
+        'char1': `איש העסקים יוסי שותף: "${topic} - חשוב לי לתרום לקהילה שלנו כי ביחד נחזקים. השקעתי השבוע ב..."
+💼 תרומה עסקית | 🤝 שיתוף קהילתי`,
+        'char2': `שרה המתנדבת מספרת: "${topic} - השבוע התנדבתי בספרייה עם הילדים ומה שקרה פה היה פשוט קסום..."
+👩‍👧‍👦 אמא מתנדבת | ✨ יצירה וחינוך`,
+        'char3': `עמותת 'יד ביד' מעדכנת: "${topic} - הארגנו השבוע אירוע קהילתי נהדר! תודה לכל המתנדבים..."
+🏢 עמותה קהילתית | 🤲 עזרה הדדית`,
+        'char4': `דני הסטודנט שותף: "${topic} - כמו סטודנט שמתמחה בתכנות, רציתי לשתף איתכם..."
+💻 סטודנט טכנולוגיה | 🚗 טרמפים`,
+        'char5': `רחל, אמא חד הורית מודה: "${topic} - כאמא לשתיים, הקהילה הזאת מאפשרת לי לתת ולקבל..."
+👩‍👧‍👧 אמא חד הורית | 💪 חוזק קהילתי`,
+        'char6': `משה הפרילנסר מציע: "${topic} - כמעצב גרפי, אני מאמין בכוח של עיצוב טוב לשנות..."
+🎨 מעצב גרפי | 💡 יצירתיות`,
+        'char7': `ליאת הקשישה הפעילה מלמדת: "${topic} - בגיל שלי למדתי שחכמת החיים היא לתת ולקבל..."
+👵 קשישה פעילה | 🧶 סריגה ותרבות`
+      };
+      return descriptions[character.id as keyof typeof descriptions] || 
+        `${character.name} שותף: "${topic} - הקהילה שלנו היא המקום שבו אני מרגיש שייך ויכול לתרום..."\n🌟 חבר קהילה | 💫 תרומה משמעותית`;
+    };
     
     data.push({
       id: `${type}-${i}`,
       type,
-      title: `${type === 'post' ? 'פוסט' : 'ריל'} #${i}`,
-      description: `תיאור מעניין של ${type === 'post' ? 'הפוסט' : 'הריל'} מספר ${i}. זהו תוכן קהילתי שמעודד שיתוף ודיון.`,
+      title: `${randomTopic} | ${randomCharacter.name}`,
+      description: getCharacterSpecificDescription(randomCharacter, randomTopic),
       thumbnail: `https://picsum.photos/seed/${type}-${i}/300/200`,
       user: {
-        id: `user-${i}`,
-        name: randomUser,
-        avatar: `https://picsum.photos/seed/user-${i}/100/100`,
-        karmaPoints: Math.floor(Math.random() * 1000) + 100,
+        id: randomCharacter.id,
+        name: randomCharacter.name,
+        avatar: randomCharacter.avatar,
+        karmaPoints: randomCharacter.karmaPoints,
       },
       likes: randomLikes,
       comments: randomComments,
-      isLiked: Math.random() < 0.3, // 30% סיכוי שהמשתמש הנוכחי עשה לייק
-      timestamp: `${Math.floor(Math.random() * 24)} שעות`,
+      isLiked: Math.random() < 0.3, // 30% chance current user liked it
+      timestamp: `${Math.floor(Math.random() * 72) + 1} שעות`,
     });
   }
   return data;
@@ -103,10 +134,13 @@ const PostReelItem = ({ item }: { item: Item }) => {
   };
 
   const handleProfilePress = () => {
-    // ניווט לפרופיל היוזר
+    // Navigate to user profile with character data
+    console.log(`🔗 Navigating to profile: ${item.user.name} (${item.user.id})`);
     (navigation as any).navigate('UserProfileScreen', { 
       userId: item.user.id,
-      userName: item.user.name 
+      userName: item.user.name,
+      // Pass additional character data for better profile display
+      characterData: characterTypes.find(char => char.id === item.user.id)
     });
   };
 
@@ -316,10 +350,14 @@ const styles = StyleSheet.create({
     overflow: 'hidden',
     backgroundColor: colors.backgroundPrimary,
     elevation: 2,
-    shadowColor: colors.shadowLight,
-    shadowOpacity: 0.1,
-    shadowOffset: { width: 0, height: 1 },
-    shadowRadius: 4,
+    ...(Platform.OS === 'web' ? {
+      boxShadow: '0 1px 4px rgba(0, 0, 0, 0.1)'
+    } : {
+      shadowColor: colors.shadowLight,
+      shadowOpacity: 0.1,
+      shadowOffset: { width: 0, height: 1 },
+      shadowRadius: 4,
+    }),
   },
   reelItem: {
     backgroundColor: '#e0f7fa', // צבע שונה לרילס
