+++ components/TimeInput.tsx
@@ -1,6 +1,6 @@
 // components/TimeInput.tsx
 import React, { useEffect, useState } from "react";
-import { UI_TEXT } from '../globals/constants';
+import { texts } from '../globals/texts';
 import {
   View,
   Text,
@@ -33,7 +33,7 @@ export default function TimeInput({
   value,
   onChange,
   label,
-  placeholder = UI_TEXT.timePlaceholder,
+  placeholder = texts.timePlaceholder,
 }: TimeInputProps) {
   const [time, setTime] = useState<Date | null>(value || null);
   const [showPicker, setShowPicker] = useState(false);
