+++ globals/constants.tsx
@@ -1,213 +1,7 @@
-export const charityNames: string[] = [
-    "לב זהב - עמותה לתמיכה בקשישים",
-    "אור לילדים - עמותה לקידום חינוך",
-    "יד למשפחה - סיוע למשפחות במצוקה",
-    "חמלה לבעלי חיים - עמותה להצלת חיות",
-    "פעמוני תקווה - תמיכה בנוער בסיכון",
-    "זרעים של שלום - קידום דו-קיום",
-    "כנפיים של חופש - סיוע לנפגעי אלימות",
-    "עתיד טוב יותר - עמותה לפיתוח קהילתי",
-    "קול התקווה - תמיכה בחולי סרטן",
-    "שביל האור - ליווי אנשים עם מוגבלויות",
-    "בריאות לכולם - קידום רפואה נגישה",
-    "ירוק בעיניים - שמירה על איכות הסביבה",
-    "דרך הלב - סיוע פסיכולוגי",
-    "גשר לאקדמיה - קידום השכלה גבוהה",
-    "בית חם לניצולים - תמיכה בניצולי שואה",
-    "לגעת בשמיים - סיוע לילדים חולים",
-    "נתינה לקהילה - פרויקטים חברתיים",
-    "מרפא לנשמה - טיפול בטראומה",
-    "עוגן בים - תמיכה באוכלוסיות ימיות",
-    "יד רחבה - חלוקת מזון לנזקקים",
-    "קשת בענן - תמיכה בקהילת הלהטב",
-    "למען הזולת - התנדבות ועזרה הדדית",
-    "אומנות למען שינוי - קידום חברתי דרך אומנות",
-    "שוויון הזדמנויות - עמותה לצדק חברתי",
-    "מגן לטף - הגנה על ילדים בסיכון",
-    "פנים חדשות - שיקום אסירים",
-    "חיים בכבוד - סיוע לחסרי בית",
-    "קול הנשים - קידום זכויות נשים",
-    "למען הטבע - שימור ביו-מגוון",
-    "דרך היצירה - העצמה באמצעות יצירה",
-    "הבטחון התזונתי - מניעת רעב",
-    "מעגלי חיים - תמיכה במשפחות שכולות",
-    "רשת הביטחון - סיוע לעולים חדשים",
-    "אורח חיים בריא - קידום בריאות הציבור",
-    "מטיילים למען הקהילה - עמותת טיולים אתגריים",
-    "שקט נפשי - סיוע נפשי לנוער",
-    "התפתחות אישית - סדנאות וקורסים",
-    "רוח ספורטיבית - קידום ספורט תחרותי",
-    "מסע אל העצמי - עמותה לפיתוח מודעות",
-    "בונים גשרים - חיבור בין קהילות",
-    "אהבת הארץ - שימור מורשת",
-    "קול אחד - עמותה לזכויות אדם",
-    "חיוך לכולם - טיפול שיניים ללא עלות",
-    "מלאכים בלבן - תמיכה במשפחות של חולים כרוניים",
-    "הזדמנות שנייה - שיקום חברתי",
-    "דרך השלום - חינוך לסובלנות",
-    "אמץ חבר - עמותה לאימוץ כלבים וחתולים",
-    "יד תומכת - ליווי יולדות",
-    "מפגש בין דורי - חיבור בין ותיקים לצעירים",
-    "עושים טוב - חלוקת ציוד לבית הספר",
-    "קריאת כיוון - ייעוץ תעסוקתי",
-    "עולם חדש - תמיכה בילדים עם צרכים מיוחדים",
-    "ללא גבולות - עמותה לקידום ספורט נכים",
-    "ניצוץ של תקווה - עמותה לסיוע לנפגעי סמים",
-    "דרך היין - קידום תרבות היין המקומית",
-    "כלים לחיים - סדנאות מיומנויות",
-    "פרנסה בכבוד - סיוע בהשמה לעבודה",
-    "שומרי היער - עמותה למען העצים",
-    "תקווה לשינוי - טיפול בהתמכרויות",
-    "מעוף הציפור - סיוע נפשי לילדים",
-    "הכיפה הסגולה - עמותה לטיפול באוטיזם",
-    "פני מלאך - סיוע למשפחות עם ילדים חולים",
-    "שער הזהב - עמותה לגמלאים",
-    "קול הדממה - סיוע ללקויי שמיעה",
-    "הדרך הטובה - קידום ערכים",
-    "לב פתוח - עמותה לתרומת איברים",
-    "צבעוניות בחיים - תמיכה באומנים מתחילים",
-    "כוח המחשבה - עמותה למען חולי אלצהיימר",
-    "עולם שלם - שילוב עולים בחברה",
-    "ניצוצות של אור - סיוע לילדים בסיכון",
-    "בטחה לילד - הגנה מפני התעללות",
-    "פעמוני חירות - סיוע לאסירים משוחררים",
-    "מזון לכולם - בנק מזון ארצי",
-    "קול הארץ - שימור אתרי מורשת",
-    "לחיים טובים יותר - עמותה לבריאות הנפש",
-    "אור השחר - סיוע לנפגעי כתות",
-    "השראה לצעירים - פיתוח מנהיגות",
-    "יד ביד לקהילה - פרויקטים קהילתיים",
-    "בונים עתיד - שיקום שכונות מצוקה",
-    "קול השקט - עמותה למען נפגעי טרור",
-    "דרך המעשה - קידום התנדבות",
-    "גן עדן לחיות - מקלט לבעלי חיים",
-    "שחף של חופש - סיוע לנוער בסיכון",
-    "לב האדם - קידום צדק חברתי",
-    "כחול לבן - עמותה למען חיילים בודדים",
-    "התקווה שבלב - סיוע למשפחות פשע",
-    "אור בחשיכה - תמיכה נפשית",
-    "למען הדור הבא - חינוך סביבתי",
-    "קול העם - עמותה לזכויות צרכנים",
-    "מרכז האושר - סדנאות לרווחה אישית",
-    "חזון ויוזמה - קידום יזמות חברתית",
-    "מפתחות להצלחה - סיוע בחינוך מקצועי",
-    "שלום עולמי - קידום דיאלוג בין-תרבותי",
-    "בית לגמלאי - פעילויות לגיל הזהב",
-    "תרומה לחיים - קידום תרומת דם",
-    "מעגל הנתינה - סיוע לנזקקים בחגים",
-    "דרך החינוך - קידום ערכים בבתי ספר",
-    "אוצרות הטבע - שימור אוצרות טבע",
-    "קול המצוקה - סיוע לנפגעי אלימות במשפחה",
-    "שדות ירוקים - חקלאות בת קיימא",
-    "כנף רחבה - סיוע לילדים עם מחלות נדירות",
-    "יד ביד למחר - קידום שילוב בחברה",
-    "שקט ורוגע - מרכז לטיפול בסטרס",
-    "מחשבה ירוקה - קידום מודעות סביבתית",
-    "הלב הפועם - סיוע למשפחות עם ילדים חולי לב",
-    "מסע של אור - עמותה לליווי מטופלים אונקולוגיים"
-  ];
+// Charities list moved to fakeData.ts file
 
 
-export const motivationalQuotes = [
-  "העתיד שייך לאלו שמאמינים ביופיים של חלומותיהם.",
-  "כל יום הוא התחלה חדשה.",
-  "אל תוותר – ההתחלה תמיד הכי קשה.",
-  "הדרך היחידה לעשות עבודה נהדרת היא לאהוב את מה שאתה עושה.",
-  "אין דבר העומד בפני הרצון.",
-  "כשאתה מפחד – תזוז בכל זאת.",
-  "כל כישלון הוא צעד אחד לקראת ההצלחה.",
-  "אל תשווה את ההתחלה שלך לאמצע של מישהו אחר.",
-  "עשה היום את מה שאחרים לא – כדי שתוכל מחר מה שאחרים לא יכולים.",
-  "המכשולים הם מה שאתה רואה כשאתה מוריד את העיניים מהמטרה.",
-  "האומץ לא תמיד שואג – לפעמים הוא לוחש: אנסה שוב מחר.",
-  "לא משנה כמה לאט אתה מתקדם – אתה עדיין עוקף את מי שיושב על הספה.",
-  "כל יום הוא הזדמנות להתקדם.",
-  "הצלחה היא סדרה של צעדים קטנים שנעשים בעקביות.",
-  "אין קיצורי דרך למקומות ששווה להגיע אליהם.",
-  "תאמין בעצמך גם כשאף אחד אחר לא.",
-  "כל מסע מתחיל בצעד אחד.",
-  "הזמן הכי טוב להתחיל היה אתמול. השני הכי טוב הוא עכשיו.",
-  "אם אתה יכול לחלום את זה – אתה יכול להשיג את זה.",
-  "אל תיתן לפחד להחליט בשבילך.",
-  "אתה חזק יותר ממה שאתה חושב.",
-  "הצלחה מתחילה בהחלטה לנסות.",
-  "אל תחכה להזדמנות – צור אותה.",
-  "בכל נפילה טמונה צמיחה.",
-  "גם הדרך הארוכה ביותר מתחילה בצעד קטן.",
-  "מה שלא הורג – מחשל.",
-  "כדי לנצח – צריך קודם להאמין שאפשר.",
-  "תחלום בגדול – תעבוד חכם – תישאר צנוע.",
-  "הכוח האמיתי נמצא בהתמדה.",
-  "להיות אמיץ זה להתחיל גם כשאין ביטחון.",
-  "השינוי מתחיל מבפנים.",
-  "מה שאתה מאמין – זה מה שאתה יוצר.",
-  "תהיה השינוי שאתה רוצה לראות בעולם.",
-  "אל תוותר על עצמך – אף פעם.",
-  "העבר לא שווה להישאר בו – ההווה מחכה שתיצור אותו.",
-  "כשיש למה לקום – הקימה קלה יותר.",
-  "חולשה היא רק חוסר תרגול.",
-  "כל קושי הוא שיעור במסע.",
-  "בכל יום אתה נולד מחדש – מה תעשה היום?",
-  "החלום שלך שווה את זה.",
-  "אם תיפול 7 פעמים – קום 8.",
-  "תודה היא המפתח לשפע.",
-  "החיים קצרים מדי כדי לדחות את החלומות.",
-  "תהיה טוב – גם כשאף אחד לא רואה.",
-  "אל תחפש השראה – תהיה השראה.",
-  "האמת שלך היא העוצמה שלך.",
-  "הדרך הקצרה היא להמשיך.",
-  "לכל אתגר יש פתרון – לפעמים אתה הפתרון.",
-  "מה שאתה מתרכז בו – גדל.",
-  "אל תתנצל על זה שאתה משתנה.",
-  "אם אין דרך – תיצור אחת.",
-  "תמשיך לזוז – גם אם לאט.",
-  "תהיה עסוק בלבנות – לא בלבקר.",
-  "האור תמיד חזק מהחושך.",
-  "העולם מחכה לגרסה הכי טובה שלך.",
-  "כל אדם שאתה פוגש – נלחם בקרב שאתה לא רואה.",
-  "פשוט תתחיל.",
-  "תן למה שמדליק אותך – לשרוף את הספקות.",
-  "עשה את הכי טוב שאתה יכול – במקום שאתה נמצא.",
-  "הסבלנות היא חלק מההצלחה.",
-  "אל תפחד לזרוח.",
-  "כל כישרון צריך תרגול – כל חלום צריך אומץ.",
-  "אל תיתן לפחד להשתיק אותך.",
-  "תצא מהקופסה – או תבנה אחת חדשה.",
-  "כל יום – הזדמנות חדשה להתחזק.",
-  "גם דרך קשה מובילה לפסגות גבוהות.",
-  "כאב זמני – גאווה לנצח.",
-  "החלומות שלך לא במקרה שלך.",
-  "תאמין, תפעל, תחזור על זה.",
-  "הדרך להצלחה רצופה בניסיונות.",
-  "ההתמדה מנצחת את הכישרון – אם הכישרון לא מתמיד.",
-  "אל תחכה לרגע הנכון – תיצור אותו.",
-  "אין הגבלה למה שאתה יכול להיות.",
-  "תאהב את הדרך – לא רק את היעד.",
-  "מי שנכשל – ניסה. מי שלא – כבר נכשל.",
-  "הניצחון האמיתי הוא פנימי.",
-  "המפתח הוא לא לוותר.",
-  "תהיה עקשן לגבי החלומות שלך – אבל גמיש בדרך.",
-  "מה שאתה עושה – חשוב יותר ממה שאתה אומר.",
-  "אל תיתן לקושי לעצור אותך – תן לו לדחוף אותך.",
-  "תזכור למה התחלת.",
-  "תשתדל להיות טוב יותר מהאתמול – זה מספיק.",
-  "כל שינוי גדול – התחיל בצעד אמיץ קטן.",
-  "אל תסתכל אחורה – אתה לא הולך לשם.",
-  "האומץ מתחיל בהחלטה.",
-  "כשאתה נופל – זו רק הזדמנות לקום חזק יותר.",
-  "אל תיתן לפחד לגנוב לך את החלום.",
-  "תבחר באור – כל יום מחדש.",
-  "הדרך הכי טובה לנבא את העתיד – היא ליצור אותו.",
-  "גם צעד קטן קדימה – זה עדיין קדימה.",
-  "תהיה הסיבה שמישהו מחייך היום.",
-  "הזמן עובר – תשתמש בו נכון.",
-  "כל יום שאתה בוחר לקום – זה ניצחון.",
-  "אמונה בעצמך – היא תחילתו של כל שינוי.",
-  "התקדמות אמיתית נמדדת בהתמדה – לא במהירות.",
-  "אין דבר חזק יותר מלב בוער.",
-  "בכל פעם שאתה מוותר – מישהו אחר מנצח.",
-  "תאהב את עצמך – אתה התחלה של כל השאר.",
-] 
+// Motivational quotes moved to fakeData.ts file
 
 export const FontSizes = {
     // --- General Text Sizes ---
@@ -289,343 +83,165 @@ export const menu_for_trumps: string[] = [
   'משלוחים',  
 ]
 
-export const WHATSAPP_GROUP_DETAILS = [
-  {
-    name: "טרמפים מרכז",
-    link: "https://chat.whatsapp.com/0lLT8M8RkPILPAV9IPfpjT",
-    image: require("../assets/images/logo.png"),
+// WhatsApp groups details moved to fakeData.ts file
+// All texts moved to texts.ts file
+
+// Bubble Component Constants
+export const BUBBLE_CONSTANTS = {
+  NUM_BUBBLES: 30,
+  MIN_SIZE: 50,
+  MAX_SIZE: 150,
+  MIN_NUMBER: 1,
+  MAX_NUMBER: 100,
+  ANIMATION_DURATION: 4000,
+  ANIMATION_DELAY: 2000,
+  OVERLAP_THRESHOLD: 2.5,
+  MAX_ATTEMPTS: 1000,
+  HEIGHT_OFFSET: 150,
+  FONT_SIZE_RATIO: 7,
+  NAME_SIZE_RATIO: 10,
+  MIN_FONT_SIZE: 10,
+  MIN_NAME_SIZE: 8,
+  TEXT_SIZE_RATIO: 0.2,
+  NAME_LINE_HEIGHT_RATIO: 1.2,
+  MIN_FONT_SCALE: 0.5,
+  MIN_NAME_SCALE: 0.3,
+  MOVEMENT_BACKGROUND: 8,
+  MOVEMENT_FOREGROUND: 5,
+  OPACITY_BACKGROUND: 0.2,
+  OPACITY_DEFAULT: 0.7,
+  OPACITY_SELECTED: 1,
+  SCALE_SELECTED: 1.1,
+  SCALE_DEFAULT: 1,
+  ANIMATION_DAMPING: 10,
+  ANIMATION_STIFFNESS: 100,
+  ANIMATION_DURATION_SHORT: 200,
+  SHADOW_OPACITY: 0.1,
+  SHADOW_OFFSET: -1,
+  SHADOW_RADIUS: 3,
+  MESSAGE_MARGIN_VERTICAL: 80,
+  MESSAGE_MARGIN_HORIZONTAL: 20,
+  MESSAGE_PADDING_HORIZONTAL: 5,
+  MESSAGE_PADDING_VERTICAL: 8,
+  MESSAGE_BORDER_RADIUS: 20,
+  MESSAGE_SHADOW_OFFSET: 2,
+  MESSAGE_SHADOW_OPACITY: 0.25,
+  MESSAGE_SHADOW_RADIUS: 4,
+  MESSAGE_OPACITY: 0.78,
+  MESSAGE_ELEVATION: 1,
+  TITLE_MARGIN_TOP: 10,
+  TITLE_FONT_SIZE: 25,
+};
+
+// Animation constants
+export const ANIMATION_CONSTANTS = {
+  SPRING_CONFIG: {
+    damping: 10,
+    stiffness: 100,
   },
-  {
-    name: "טרמפים צפון",
-    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
-    image: require("../assets/images/logo.png"),
+  TIMING_CONFIG: {
+    duration: 300,
   },
-  {
-    name: "טרמפים צפון",
-    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
-    image: require("../assets/images/logo.png"),
+  BUBBLE_MOVEMENT: {
+    SLOW: 2000,
+    MEDIUM: 3000,
+    FAST: 4000,
   },
-  {
-    name: "טרמפים צפון",
-    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
-    image: require("../assets/images/logo.png"),
+};
+
+// Screen dimensions and layout constants
+export const LAYOUT_CONSTANTS = {
+  PANEL_OFFSET: 50,
+  BOTTOM_NAV_HEIGHT: 80,
+  TOP_BAR_HEIGHT: 60,
+  BORDER_RADIUS: {
+    SMALL: 8,
+    MEDIUM: 15,
+    LARGE: 20,
+    XLARGE: 25,
   },
-  {
-    name: "טרמפים צפון",
-    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
-    image: require("../assets/images/logo.png"),
+  SPACING: {
+    XS: 4,
+    SM: 8,
+    MD: 16,
+    LG: 24,
+    XL: 32,
   },
-  {
-    name: "טרמפים צפון",
-    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
-    image: require("../assets/images/logo.png"),
+  SHADOW: {
+    LIGHT: {
+      shadowOffset: { width: 0, height: 2 },
+      shadowOpacity: 0.1,
+      shadowRadius: 3,
+      elevation: 3,
+    },
+    MEDIUM: {
+      shadowOffset: { width: 0, height: 4 },
+      shadowOpacity: 0.2,
+      shadowRadius: 5,
+      elevation: 5,
+    },
+    HEAVY: {
+      shadowOffset: { width: 0, height: 8 },
+      shadowOpacity: 0.3,
+      shadowRadius: 10,
+      elevation: 10,
+    },
   },
-  {
-    name: "טרמפים צפון",
-    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
-    image: require("../assets/images/logo.png"),
-  },
-  {
-    name: "טרמפים צפון",
-    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
-    image: require("../assets/images/logo.png"),
-  },
-  {
-    name: "טרמפים דרום",
-    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
-    image: require("../assets/images/logo.png"),
-  }, // Assuming a different name for the duplicated link
+};
+
+// Categories for tasks and donations
+export const categories = [
+  'מזון', 'בגדים', 'חינוך', 'בריאות', 'סביבה', 'קשישים', 'ילדים', 'חיות', 'תרבות', 'ספורט', 'ייעוץ', 'תחבורה'
 ];
 
-// UI Text Constants
-export const UI_TEXT = {
-  // Common Actions
-  back: "חזור",
-  save: "שמור",
-  delete: "מחק",
-  edit: "ערוך",
-  add: "הוסף",
-  search: "חיפוש",
-  filter: "סינון",
-  sort: "מיון",
-  confirm: "אישור",
-  close: "סגור",
-  next: "הבא",
-  previous: "הקודם",
-  submit: "שלח",
-  loading: "טוען...",
-  error: "שגיאה",
-  success: "הצלחה",
-  warning: "אזהרה",
-  info: "מידע",
+// Task priorities
+export const priorities = [
+  { value: 'low', label: 'נמוכה', color: '#4CAF50' },
+  { value: 'medium', label: 'בינונית', color: '#FF9800' },
+  { value: 'high', label: 'גבוהה', color: '#F44336' }
+];
 
-  // Placeholders
-  searchPlaceholder: "חפש...",
-  selectPlaceholder: "בחר...",
-  timePlaceholder: "בחר שעה",
-  locationPlaceholder: "בחר מיקום",
-  namePlaceholder: "הכנס שם",
-  emailPlaceholder: "הכנס אימייל",
-  phonePlaceholder: "הכנס טלפון",
-  messagePlaceholder: "הכנס הודעה",
+// Task statuses
+export const statuses = [
+  { value: 'pending', label: 'ממתין', color: '#FF9800' },
+  { value: 'in_progress', label: 'בביצוע', color: '#2196F3' },
+  { value: 'completed', label: 'הושלם', color: '#4CAF50' },
+  { value: 'cancelled', label: 'בוטל', color: '#9E9E9E' }
+];
 
-  // Labels
-  timeLabel: "שעה",
-  dateLabel: "תאריך",
-  locationLabel: "מיקום",
-  nameLabel: "שם",
-  emailLabel: "אימייל",
-  phoneLabel: "טלפון",
-  messageLabel: "הודעה",
-  amountLabel: "סכום",
-  categoryLabel: "קטגוריה",
-  statusLabel: "סטטוס",
+// Component size constants
+export const COMPONENT_SIZES = {
+  AVATAR: {
+    SMALL: 30,
+    MEDIUM: 50,
+    LARGE: 80,
+    XLARGE: 120,
+  },
+  BUTTON_HEIGHT: {
+    SMALL: 32,
+    MEDIUM: 44,
+    LARGE: 56,
+  },
+  INPUT_HEIGHT: {
+    SMALL: 36,
+    MEDIUM: 48,
+    LARGE: 56,
+  },
+};
 
-  // Status Messages
-  completed: "הושלם",
-  pending: "ממתין",
-  active: "פעיל",
-  inactive: "לא פעיל",
-  available: "זמין",
-  unavailable: "לא זמין",
-  online: "מקוון",
-  offline: "לא מקוון",
+// API and network constants
+export const API_CONSTANTS = {
+  REQUEST_TIMEOUT: 10000,
+  RETRY_ATTEMPTS: 3,
+  RETRY_DELAY: 1000,
+};
 
-  // Error Messages
-  searchTextRequired: "אנא הכנס טקסט לחיפוש.",
-  networkError: "שגיאת רשת",
-  generalError: "אירעה שגיאה",
-  tryAgain: "נסה שוב",
-
-  // Screen Titles
-  aboutTitle: "אודות",
-  donationsTitle: "תרומות",
-  moneyTitle: "תרומת כסף",
-  timeTitle: "תרומת זמן",
-  knowledgeTitle: "תרומת ידע",
-  trumpTitle: "טרמפים",
-  profileTitle: "פרופיל",
-  searchTitle: "חיפוש",
-  homeTitle: "בית",
-  chatTitle: "צ'אטים",
-
-  // Quick Actions
-  urgentTasks: "משימות דחופות",
-  upcomingEvents: "אירועים קרובים",
-  newDonations: "תרומות חדשות",
-  activeVolunteers: "מתנדבים פעילים",
-
-  // Community Stats
-  activeMembers: "חברים פעילים",
-  monthlyDonations: "תרומות החודש",
-  monthlyGrowth: "גידול חודשי",
-  activeCities: "ערים פעילות",
-
-  // Inactive Screen
-  inactiveMessage: "השימוש במסך זה עדיין אינו זמין. תודה על ההבנה!",
-
-  // About Screen
-  aboutSubtitle: "הקיבוץ הקפיטליסטי הראשון מסוגו",
-  communityTitle: "אודות קהילת קארמה (Karma Community)",
-  whatIsTitle: "מהי קהילת קארמה?",
-  aspirationsTitle: "השאיפות המרכזיות שלנו",
-  challengesTitle: "האתגרים הקיימים כיום (הבעיות שאנו פותרים)",
-  visionTitle: "החזון והפתרונות שאנו מציעים",
-  duplicatesTitle: "כפילויות",
-  reliabilityTitle: "חוסר אמינות",
-  dispersionTitle: "פיזור",
-
-  // Categories
-  education: "חינוך",
-  health: "בריאות",
-  welfare: "רווחה",
-  environment: "סביבה",
-  animals: "בעלי חיים",
-  youthAtRisk: "נוער בסיכון",
-  elderly: "קשישים",
-  disabled: "נכים",
-  patients: "חולים",
-  familiesInNeed: "משפחות במצוקה",
-  newImmigrants: "עולים חדשים",
-  community: "קהילה",
-  work: "עבודה",
-  studies: "לימודים",
-  leisure: "בילוי",
-
-  // Cities
-  telAviv: "תל אביב",
-  jerusalem: "ירושלים",
-  haifa: "חיפה",
-  beerSheva: "באר שבע",
-  eilat: "אילת",
-  ashdod: "אשדוד",
-  ramatGan: "רמת גן",
-  petahTikva: "פתח תקווה",
-
-  // Sort Options
-  alphabetical: "אלפביתי",
-  byLocation: "לפי מיקום",
-  byCategory: "לפי תחום",
-  byPrice: "לפי מחיר",
-  byTime: "לפי שעה",
-  byRating: "לפי דירוג",
-  byRelevance: "לפי רלוונטיות",
-  byEstablishmentDate: "לפי תאריך הקמה",
-  byDonorsCount: "לפי מספר תורמים",
-
-  // Trump Specific
-  departureTime: "שעת יציאה",
-  selectDepartureTime: "בחר שעת יציאה",
-  searchAvailableTrumps: "חפש טרמפים זמינים",
-  oneWay: "בחד יעד",
-  searchCharitiesForDonation: "חפש עמותות לתרומה",
-  searchCharitiesForHelp: "חפש עמותות שיכולות לעזור לך מבחינה כספית",
-  searchCoursesAndLessons: "חפש קורסים ושיעורים",
-
-  // WhatsApp Groups
-  trumpsCenter: "טרמפים מרכז",
-  trumpsNorth: "טרמפים צפון",
-  trumpsSouth: "טרמפים דרום",
-
-  // Group Types
-  whatsapp: "whatsapp",
-  facebook: "facebook",
-
-  // Driver Names (for demo data)
-  davidCohen: "דוד כהן",
-  saraLevy: "שרה לוי",
-  mosheGoldberg: "משה גולדברג",
-  rachelAbraham: "רחל אברהם",
-  yossiShapira: "יוסי שפירא",
-  michalRosen: "מיכל רוזן",
-
-  // Organization Names (for demo data)
-  goldenHeart: "לב זהב - עמותה לתמיכה בקשישים",
-  lightForChildren: "אור לילדים - עמותה לקידום חינוך",
-  compassionForAnimals: "חמלה לבעלי חיים - עמותה להצלת חיות",
-  healthForAll: "בריאות לכולם - קידום רפואה נגישה",
-  greenEyes: "ירוק בעיניים - שמירה על איכות הסביבה",
-  bellsOfHope: "פעמוני תקווה - תמיכה בנוער בסיכון",
-  pathOfLight: "שביל האור - ליווי אנשים עם מוגבלויות",
-  voiceOfHope: "קול התקווה - תמיכה בחולי סרטן",
-  warmHome: 'בית אבות "בית חם"',
-  goldenHeartShelter: 'מקלט חיות "לב זהב"',
-  gatesOfRighteousness: 'בית חולים "שערי צדק"',
-
-  // Descriptions (for demo data)
-  elderlySupportDesc: "תמיכה בקשישים בודדים וסיוע יומיומי",
-  educationPromotionDesc: "קידום חינוך לילדים ממשפחות מעוטות יכולת",
-  animalRescueDesc: "הצלה וטיפול בבעלי חיים נטושים",
-  accessibleMedicineDesc: "קידום רפואה נגישה לכל האוכלוסיות",
-  environmentalProtectionDesc: "שמירה על איכות הסביבה והטבע",
-  youthAtRiskSupportDesc: "תמיכה וטיפול בנוער בסיכון",
-  disabledSupportDesc: "ליווי ושילוב אנשים עם מוגבלויות",
-  cancerSupportDesc: "תמיכה נפשית ופיזית בחולי סרטן",
-
-  // Search Results
-  noResultsFound: "לא נמצאו תוצאות",
-  tryChangingSearchTerms: "נסה לשנות את מילות החיפוש או לבדוק את האיות",
-  searching: "מחפש...",
-  resultsFound: "נמצאו {count} תוצאות עבור \"{query}\"",
-  noOptions: "אין תוצאות",
-
-  // Section Titles
-  realTimeData: "נתונים בזמן אמת",
-  communityImpact: "השפעה קהילתית",
-  topContributors: "תורמים מובילים החודש",
-  weeklyActivity: "פעילות שבועית",
-  quickActions: "פעולות מהירות",
-  activityGraphPlaceholder: "גרף פעילות יופיע כאן",
-
-  // Impact Stats
-  sharedRides: "נסיעות שותפו החודש",
-  donatedMeals: "ארוחות נתרמו",
-  mentoringHours: "שעות חונכות",
-  supportedFamilies: "משפחות נתמכו",
-
-  // Bubble Component Constants
-  BUBBLE_NUM_BUBBLES: 30,
-  BUBBLE_MIN_SIZE: 50,
-  BUBBLE_MAX_SIZE: 150,
-  BUBBLE_MIN_NUMBER: 1,
-  BUBBLE_MAX_NUMBER: 100,
-  BUBBLE_ANIMATION_DURATION: 4000,
-  BUBBLE_ANIMATION_DELAY: 2000,
-  BUBBLE_OVERLAP_THRESHOLD: 2.5,
-  BUBBLE_MAX_ATTEMPTS: 1000,
-  BUBBLE_HEIGHT_OFFSET: 150,
-  BUBBLE_FONT_SIZE_RATIO: 7,
-  BUBBLE_NAME_SIZE_RATIO: 10,
-  BUBBLE_MIN_FONT_SIZE: 10,
-  BUBBLE_MIN_NAME_SIZE: 8,
-  BUBBLE_TEXT_SIZE_RATIO: 0.2,
-  BUBBLE_NAME_LINE_HEIGHT_RATIO: 1.2,
-  BUBBLE_MIN_FONT_SCALE: 0.5,
-  BUBBLE_MIN_NAME_SCALE: 0.3,
-  BUBBLE_MOVEMENT_BACKGROUND: 8,
-  BUBBLE_MOVEMENT_FOREGROUND: 5,
-  BUBBLE_OPACITY_BACKGROUND: 0.2,
-  BUBBLE_OPACITY_DEFAULT: 0.7,
-  BUBBLE_OPACITY_SELECTED: 1,
-  BUBBLE_SCALE_SELECTED: 1.1,
-  BUBBLE_SCALE_DEFAULT: 1,
-  BUBBLE_ANIMATION_DAMPING: 10,
-  BUBBLE_ANIMATION_STIFFNESS: 100,
-  BUBBLE_ANIMATION_DURATION_SHORT: 200,
-  BUBBLE_SHADOW_OPACITY: 0.1,
-  BUBBLE_SHADOW_OFFSET: -1,
-  BUBBLE_SHADOW_RADIUS: 3,
-  BUBBLE_MESSAGE_MARGIN_VERTICAL: 80,
-  BUBBLE_MESSAGE_MARGIN_HORIZONTAL: 20,
-  BUBBLE_MESSAGE_PADDING_HORIZONTAL: 5,
-  BUBBLE_MESSAGE_PADDING_VERTICAL: 8,
-  BUBBLE_MESSAGE_BORDER_RADIUS: 20,
-  BUBBLE_MESSAGE_SHADOW_OFFSET: 2,
-  BUBBLE_MESSAGE_SHADOW_OPACITY: 0.25,
-  BUBBLE_MESSAGE_SHADOW_RADIUS: 4,
-  BUBBLE_MESSAGE_OPACITY: 0.78,
-  BUBBLE_MESSAGE_ELEVATION: 1,
-  BUBBLE_TITLE_MARGIN_TOP: 10,
-  BUBBLE_TITLE_FONT_SIZE: 25,
-
-  // Bubble Component Text
-  bubbleTitle: "הקהילה במספרים",
-
-  // Settings Screen
-  settingsTitle: "הגדרות",
-  settingsSubtitle: "הגדרות האפליקציה",
-  
-  // Settings Sections
-  accountSection: "חשבון",
-  notificationsSection: "התראות", 
-  appPreferencesSection: "העדפות אפליקציה",
-  helpSupportSection: "עזרה ותמיכה",
-  
-  // Settings Items
-  editProfile: "עריכת פרופיל",
-  privacy: "פרטיות",
-  privacyVisibility: "הצג סטטוס מקוון",
-  pushNotifications: "התראות דחיפה",
-  emailNotifications: "התראות דוא\"ל",
-  darkMode: "מצב כהה",
-  dataSaver: "חיסכון בנתונים",
-  clearCache: "ניקוי מטמון",
-  helpCenter: "מרכז העזרה",
-  about: "אודות",
-  logout: "התנתק",
-  
-  // Settings Alerts
-  clearCacheTitle: "ניקוי מטמון",
-  clearCacheMessage: "האם אתה בטוח שברצונך לנקות את המטמון של האפליקציה? פעולה זו תפנה שטח אחסון.",
-  cacheClearedTitle: "המטמון נוקה",
-  cacheClearedMessage: "המטמון של האפליקציה נוקה בהצלחה.",
-  logoutTitle: "התנתק",
-  logoutMessage: "האם אתה בטוח שברצונך להתנתק?",
-  linkErrorTitle: "שגיאה",
-  linkErrorMessage: "לא ניתן לפתוח את הקישור",
-  
-  // Common Actions
-  cancel: "ביטול",
-  clear: "נקה",
+// Validation constants
+export const VALIDATION = {
+  MIN_PASSWORD_LENGTH: 6,
+  MAX_MESSAGE_LENGTH: 500,
+  MAX_TITLE_LENGTH: 100,
+  EMAIL_REGEX: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
+  PHONE_REGEX: /^[\+]?[1-9][\d]{0,15}$/,
 };
