+++ globals/fakeData.ts
@@ -638,3 +638,300 @@ export const donationStats = [
   { id: 'members1', value: 52, name: "חברים חדשים השבוע", icon: "👥", category: 'members', color: '#9C27B0' },
   { id: 'members2', value: 3247, name: "סה״כ חברי קהילה פעילים", icon: "🌟", category: 'members', color: '#BA68C8' }
 ];
+
+// === נתונים שהועברו מ-constants.tsx ===
+
+// רשימת שמות עמותות
+export const charityNames: string[] = [
+  "לב זהב - עמותה לתמיכה בקשישים",
+  "אור לילדים - עמותה לקידום חינוך",
+  "יד למשפחה - סיוע למשפחות במצוקה",
+  "חמלה לבעלי חיים - עמותה להצלת חיות",
+  "פעמוני תקווה - תמיכה בנוער בסיכון",
+  "זרעים של שלום - קידום דו-קיום",
+  "כנפיים של חופש - סיוע לנפגעי אלימות",
+  "עתיד טוב יותר - עמותה לפיתוח קהילתי",
+  "קול התקווה - תמיכה בחולי סרטן",
+  "שביל האור - ליווי אנשים עם מוגבלויות",
+  "בריאות לכולם - קידום רפואה נגישה",
+  "ירוק בעיניים - שמירה על איכות הסביבה",
+  "דרך הלב - סיוע פסיכולוגי",
+  "גשר לאקדמיה - קידום השכלה גבוהה",
+  "בית חם לניצולים - תמיכה בניצולי שואה",
+  "לגעת בשמיים - סיוע לילדים חולים",
+  "נתינה לקהילה - פרויקטים חברתיים",
+  "מרפא לנשמה - טיפול בטראומה",
+  "עוגן בים - תמיכה באוכלוסיות ימיות",
+  "יד רחבה - חלוקת מזון לנזקקים",
+  "קשת בענן - תמיכה בקהילת הלהטב",
+  "למען הזולת - התנדבות ועזרה הדדית",
+  "אומנות למען שינוי - קידום חברתי דרך אומנות",
+  "שוויון הזדמנויות - עמותה לצדק חברתי",
+  "מגן לטף - הגנה על ילדים בסיכון",
+  "פנים חדשות - שיקום אסירים",
+  "חיים בכבוד - סיוע לחסרי בית",
+  "קול הנשים - קידום זכויות נשים",
+  "למען הטבע - שימור ביו-מגוון",
+  "דרך היצירה - העצמה באמצעות יצירה",
+  "הבטחון התזונתי - מניעת רעב",
+  "מעגלי חיים - תמיכה במשפחות שכולות",
+  "רשת הביטחון - סיוע לעולים חדשים",
+  "אורח חיים בריא - קידום בריאות הציבור",
+  "מטיילים למען הקהילה - עמותת טיולים אתגריים",
+  "שקט נפשי - סיוע נפשי לנוער",
+  "התפתחות אישית - סדנאות וקורסים",
+  "רוח ספורטיבית - קידום ספורט תחרותי",
+  "מסע אל העצמי - עמותה לפיתוח מודעות",
+  "בונים גשרים - חיבור בין קהילות",
+  "אהבת הארץ - שימור מורשת",
+  "קול אחד - עמותה לזכויות אדם",
+  "חיוך לכולם - טיפול שיניים ללא עלות",
+  "מלאכים בלבן - תמיכה במשפחות של חולים כרוניים",
+  "הזדמנות שנייה - שיקום חברתי",
+  "דרך השלום - חינוך לסובלנות",
+  "אמץ חבר - עמותה לאימוץ כלבים וחתולים",
+  "יד תומכת - ליווי יולדות",
+  "מפגש בין דורי - חיבור בין ותיקים לצעירים",
+  "עושים טוב - חלוקת ציוד לבית הספר",
+  "קריאת כיוון - ייעוץ תעסוקתי",
+  "עולם חדש - תמיכה בילדים עם צרכים מיוחדים",
+  "ללא גבולות - עמותה לקידום ספורט נכים",
+  "ניצוץ של תקווה - עמותה לסיוע לנפגעי סמים",
+  "דרך היין - קידום תרבות היין המקומית",
+  "כלים לחיים - סדנאות מיומנויות",
+  "פרנסה בכבוד - סיוע בהשמה לעבודה",
+  "שומרי היער - עמותה למען העצים",
+  "תקווה לשינוי - טיפול בהתמכרויות",
+  "מעוף הציפור - סיוע נפשי לילדים",
+  "הכיפה הסגולה - עמותה לטיפול באוטיזם",
+  "פני מלאך - סיוע למשפחות עם ילדים חולים",
+  "שער הזהב - עמותה לגמלאים",
+  "קול הדממה - סיוע ללקויי שמיעה",
+  "הדרך הטובה - קידום ערכים",
+  "לב פתוח - עמותה לתרומת איברים",
+  "צבעוניות בחיים - תמיכה באומנים מתחילים",
+  "כוח המחשבה - עמותה למען חולי אלצהיימר",
+  "עולם שלם - שילוב עולים בחברה",
+  "ניצוצות של אור - סיוע לילדים בסיכון",
+  "בטחה לילד - הגנה מפני התעללות",
+  "פעמוני חירות - סיוע לאסירים משוחררים",
+  "מזון לכולם - בנק מזון ארצי",
+  "קול הארץ - שימור אתרי מורשת",
+  "לחיים טובים יותר - עמותה לבריאות הנפש",
+  "אור השחר - סיוע לנפגעי כתות",
+  "השראה לצעירים - פיתוח מנהיגות",
+  "יד ביד לקהילה - פרויקטים קהילתיים",
+  "בונים עתיד - שיקום שכונות מצוקה",
+  "קול השקט - עמותה למען נפגעי טרור",
+  "דרך המעשה - קידום התנדבות",
+  "גן עדן לחיות - מקלט לבעלי חיים",
+  "שחף של חופש - סיוע לנוער בסיכון",
+  "לב האדם - קידום צדק חברתי",
+  "כחול לבן - עמותה למען חיילים בודדים",
+  "התקווה שבלב - סיוע למשפחות פשע",
+  "אור בחשיכה - תמיכה נפשית",
+  "למען הדור הבא - חינוך סביבתי",
+  "קול העם - עמותה לזכויות צרכנים",
+  "מרכז האושר - סדנאות לרווחה אישית",
+  "חזון ויוזמה - קידום יזמות חברתית",
+  "מפתחות להצלחה - סיוע בחינוך מקצועי",
+  "שלום עולמי - קידום דיאלוג בין-תרבותי",
+  "בית לגמלאי - פעילויות לגיל הזהב",
+  "תרומה לחיים - קידום תרומת דם",
+  "מעגל הנתינה - סיוע לנזקקים בחגים",
+  "דרך החינוך - קידום ערכים בבתי ספר",
+  "אוצרות הטבע - שימור אוצרות טבע",
+  "קול המצוקה - סיוע לנפגעי אלימות במשפחה",
+  "שדות ירוקים - חקלאות בת קיימא",
+  "כנף רחבה - סיוע לילדים עם מחלות נדירות",
+  "יד ביד למחר - קידום שילוב בחברה",
+  "שקט ורוגע - מרכז לטיפול בסטרס",
+  "מחשבה ירוקה - קידום מודעות סביבתית",
+  "הלב הפועם - סיוע למשפחות עם ילדים חולי לב",
+  "מסע של אור - עמותה לליווי מטופלים אונקולוגיים"
+];
+
+// ציטוטי מוטיבציה
+export const motivationalQuotes = [
+  "העתיד שייך לאלו שמאמינים ביופיים של חלומותיהם.",
+  "כל יום הוא התחלה חדשה.",
+  "אל תוותר – ההתחלה תמיד הכי קשה.",
+  "הדרך היחידה לעשות עבודה נהדרת היא לאהוב את מה שאתה עושה.",
+  "אין דבר העומד בפני הרצון.",
+  "כשאתה מפחד – תזוז בכל זאת.",
+  "כל כישלון הוא צעד אחד לקראת ההצלחה.",
+  "אל תשווה את ההתחלה שלך לאמצע של מישהו אחר.",
+  "עשה היום את מה שאחרים לא – כדי שתוכל מחר מה שאחרים לא יכולים.",
+  "המכשולים הם מה שאתה רואה כשאתה מוריד את העיניים מהמטרה.",
+  "האומץ לא תמיד שואג – לפעמים הוא לוחש: אנסה שוב מחר.",
+  "לא משנה כמה לאט אתה מתקדם – אתה עדיין עוקף את מי שיושב על הספה.",
+  "כל יום הוא הזדמנות להתקדם.",
+  "הצלחה היא סדרה של צעדים קטנים שנעשים בעקביות.",
+  "אין קיצורי דרך למקומות ששווה להגיע אליהם.",
+  "תאמין בעצמך גם כשאף אחד אחר לא.",
+  "כל מסע מתחיל בצעד אחד.",
+  "הזמן הכי טוב להתחיל היה אתמול. השני הכי טוב הוא עכשיו.",
+  "אם אתה יכול לחלום את זה – אתה יכול להשיג את זה.",
+  "אל תיתן לפחד להחליט בשבילך.",
+  "אתה חזק יותר ממה שאתה חושב.",
+  "הצלחה מתחילה בהחלטה לנסות.",
+  "אל תחכה להזדמנות – צור אותה.",
+  "בכל נפילה טמונה צמיחה.",
+  "גם הדרך הארוכה ביותר מתחילה בצעד קטן.",
+  "מה שלא הורג – מחשל.",
+  "כדי לנצח – צריך קודם להאמין שאפשר.",
+  "תחלום בגדול – תעבוד חכם – תישאר צנוע.",
+  "הכוח האמיתי נמצא בהתמדה.",
+  "להיות אמיץ זה להתחיל גם כשאין ביטחון.",
+  "השינוי מתחיל מבפנים.",
+  "מה שאתה מאמין – זה מה שאתה יוצר.",
+  "תהיה השינוי שאתה רוצה לראות בעולם.",
+  "אל תוותר על עצמך – אף פעם.",
+  "העבר לא שווה להישאר בו – ההווה מחכה שתיצור אותו.",
+  "כשיש למה לקום – הקימה קלה יותר.",
+  "חולשה היא רק חוסר תרגול.",
+  "כל קושי הוא שיעור במסע.",
+  "בכל יום אתה נולד מחדש – מה תעשה היום?",
+  "החלום שלך שווה את זה.",
+  "אם תיפול 7 פעמים – קום 8.",
+  "תודה היא המפתח לשפע.",
+  "החיים קצרים מדי כדי לדחות את החלומות.",
+  "תהיה טוב – גם כשאף אחד לא רואה.",
+  "אל תחפש השראה – תהיה השראה.",
+  "האמת שלך היא העוצמה שלך.",
+  "הדרך הקצרה היא להמשיך.",
+  "לכל אתגר יש פתרון – לפעמים אתה הפתרון.",
+  "מה שאתה מתרכז בו – גדל.",
+  "אל תתנצל על זה שאתה משתנה.",
+  "אם אין דרך – תיצור אחת.",
+  "תמשיך לזוז – גם אם לאט.",
+  "תהיה עסוק בלבנות – לא בלבקר.",
+  "האור תמיד חזק מהחושך.",
+  "העולם מחכה לגרסה הכי טובה שלך.",
+  "כל אדם שאתה פוגש – נלחם בקרב שאתה לא רואה.",
+  "פשוט תתחיל.",
+  "תן למה שמדליק אותך – לשרוף את הספקות.",
+  "עשה את הכי טוב שאתה יכול – במקום שאתה נמצא.",
+  "הסבלנות היא חלק מההצלחה.",
+  "אל תפחד לזרוח.",
+  "כל כישרון צריך תרגול – כל חלום צריך אומץ.",
+  "אל תיתן לפחד להשתיק אותך.",
+  "תצא מהקופסה – או תבנה אחת חדשה.",
+  "כל יום – הזדמנות חדשה להתחזק.",
+  "גם דרך קשה מובילה לפסגות גבוהות.",
+  "כאב זמני – גאווה לנצח.",
+  "החלומות שלך לא במקרה שלך.",
+  "תאמין, תפעל, תחזור על זה.",
+  "הדרך להצלחה רצופה בניסיונות.",
+  "ההתמדה מנצחת את הכישרון – אם הכישרון לא מתמיד.",
+  "אל תחכה לרגע הנכון – תיצור אותו.",
+  "אין הגבלה למה שאתה יכול להיות.",
+  "תאהב את הדרך – לא רק את היעד.",
+  "מי שנכשל – ניסה. מי שלא – כבר נכשל.",
+  "הניצחון האמיתי הוא פנימי.",
+  "המפתח הוא לא לוותר.",
+  "תהיה עקשן לגבי החלומות שלך – אבל גמיש בדרך.",
+  "מה שאתה עושה – חשוב יותר ממה שאתה אומר.",
+  "אל תיתן לקושי לעצור אותך – תן לו לדחוף אותך.",
+  "תזכור למה התחלת.",
+  "תשתדל להיות טוב יותר מהאתמול – זה מספיק.",
+  "כל שינוי גדול – התחיל בצעד אמיץ קטן.",
+  "אל תסתכל אחורה – אתה לא הולך לשם.",
+  "האומץ מתחיל בהחלטה.",
+  "כשאתה נופל – זו רק הזדמנות לקום חזק יותר.",
+  "אל תיתן לפחד לגנוב לך את החלום.",
+  "תבחר באור – כל יום מחדש.",
+  "הדרך הכי טובה לנבא את העתיד – היא ליצור אותו.",
+  "גם צעד קטן קדימה – זה עדיין קדימה.",
+  "תהיה הסיבה שמישהו מחייך היום.",
+  "הזמן עובר – תשתמש בו נכון.",
+  "כל יום שאתה בוחר לקום – זה ניצחון.",
+  "אמונה בעצמך – היא תחילתו של כל שינוי.",
+  "התקדמות אמיתית נמדדת בהתמדה – לא במהירות.",
+  "אין דבר חזק יותר מלב בוער.",
+  "בכל פעם שאתה מוותר – מישהו אחר מנצח.",
+  "תאהב את עצמך – אתה התחלה של כל השאר.",
+  "תמיד יש איפה לגדול.",
+  "השינוי האמיתי דורש אומץ."
+];
+
+// פרטי קבוצות וואטסאפ
+export const WHATSAPP_GROUP_DETAILS = [
+  {
+    name: "טרמפים מרכז",
+    link: "https://chat.whatsapp.com/0lLT8M8RkPILPAV9IPfpjT",
+    image: require("../assets/images/logo.png"),
+  },
+  {
+    name: "טרמפים צפון",
+    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
+    image: require("../assets/images/logo.png"),
+  },
+  {
+    name: "טרמפים צפון",
+    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
+    image: require("../assets/images/logo.png"),
+  },
+  {
+    name: "טרמפים צפון",
+    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
+    image: require("../assets/images/logo.png"),
+  },
+  {
+    name: "טרמפים צפון",
+    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
+    image: require("../assets/images/logo.png"),
+  },
+  {
+    name: "טרמפים צפון",
+    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
+    image: require("../assets/images/logo.png"),
+  },
+  {
+    name: "טרמפים צפון",
+    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
+    image: require("../assets/images/logo.png"),
+  },
+  {
+    name: "טרמפים צפון",
+    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
+    image: require("../assets/images/logo.png"),
+  },
+  {
+    name: "טרמפים דרום",
+    link: "https://chat.whatsapp.com/GjHTYqHGYF63VWh3BfTbE",
+    image: require("../assets/images/logo.png"),
+  },
+];
+
+
+  // Recent Activities
+export const recentActivities = [
+    {
+      id: '1',
+      type: 'task',
+      title: 'משימה הושלמה: איסוף מזון',
+      description: 'אנה כהן השלימה משימת איסוף מזון',
+      time: 'לפני שעה',
+      icon: 'checkmark-circle',
+      color: '#4CAF50' // success green
+    },
+    {
+      id: '2',
+      type: 'donation',
+      title: 'תרומה חדשה: 500 ₪',
+      description: 'דני לוי תרם 500 ₪ לקניית ציוד',
+      time: 'לפני 3 שעות',
+      icon: 'heart',
+      color: '#F44336' // error red
+    },
+    {
+      id: '3',
+      type: 'event',
+      title: 'אירוע חדש: יום קהילה',
+      description: 'אירוע קהילתי גדול יתקיים בשבוע הבא',
+      time: 'לפני יום',
+      icon: 'calendar',
+      color: '#2196F3' // info blue
+    }
+  ];
