+++ globals/texts.ts
@@ -144,7 +144,310 @@ export const texts = {
   selectedUser: "נבחר המשתמש: {{name}}",
   postNumber: "פוסט מספר {{number}}",
   highlightNumber: "היילייט {{number}}",
-  activityTime: "{{title}} - {{time}}"
+  activityTime: "{{title}} - {{time}}",
+
+  // === טקסטים שהועברו מ-constants.tsx ===
+  
+  // Common Actions
+  back: "חזור",
+  save: "שמור",
+  delete: "מחק",
+  edit: "ערוך",
+  add: "הוסף",
+  search: "חיפוש",
+  filter: "סינון",
+  sort: "מיון",
+  confirm: "אישור",
+  close: "סגור",
+  next: "הבא",
+  previous: "הקודם",
+  submit: "שלח",
+  loading: "טוען...",
+  error: "שגיאה",
+  success: "הצלחה",
+  warning: "אזהרה",
+  info: "מידע",
+
+  // Placeholders
+  searchPlaceholder: "חפש...",
+  selectPlaceholder: "בחר...",
+  timePlaceholder: "בחר שעה",
+  locationPlaceholder: "בחר מיקום",
+  namePlaceholder: "הכנס שם",
+  emailPlaceholder: "הכנס אימייל",
+  phonePlaceholder: "הכנס טלפון",
+  messagePlaceholder: "הכנס הודעה",
+
+  // Labels
+  timeLabel: "שעה",
+  dateLabel: "תאריך",
+  locationLabel: "מיקום",
+  nameLabel: "שם",
+  emailLabel: "אימייל",
+  phoneLabel: "טלפון",
+  messageLabel: "הודעה",
+  amountLabel: "סכום",
+  categoryLabel: "קטגוריה",
+  statusLabel: "סטטוס",
+
+  // Status Messages
+  completed: "הושלם",
+  pending: "ממתין",
+  active: "פעיל",
+  inactive: "לא פעיל",
+  available: "זמין",
+  unavailable: "לא זמין",
+  online: "מקוון",
+  offline: "לא מקוון",
+
+  // Error Messages
+  searchTextRequired: "אנא הכנס טקסט לחיפוש.",
+  networkError: "שגיאת רשת",
+  generalError: "אירעה שגיאה",
+  tryAgain: "נסה שוב",
+
+  // Screen Titles
+  aboutTitle: "אודות",
+  donationsTitle: "תרומות",
+  moneyTitle: "תרומת כסף",
+  timeTitle: "תרומת זמן",
+  knowledgeTitle: "תרומת ידע",
+  trumpTitle: "טרמפים",
+  profileTitle: "פרופיל",
+  searchTitle: "חיפוש",
+  homeTitle: "בית",
+  chatTitle: "צ'אטים",
+
+  // Quick Actions
+  urgentTasks: "משימות דחופות",
+  upcomingEvents: "אירועים קרובים",
+  newDonations: "תרומות חדשות",
+  activeVolunteers: "מתנדבים פעילים",
+
+  // Community Stats
+  activeMembers: "חברים פעילים",
+  monthlyDonations: "תרומות החודש",
+  monthlyGrowth: "גידול חודשי",
+  activeCities: "ערים פעילות",
+
+  // Inactive Screen
+  inactiveMessage: "השימוש במסך זה עדיין אינו זמין. תודה על ההבנה!",
+
+  // About Screen
+  aboutSubtitle: "הקיבוץ הקפיטליסטי הראשון מסוגו",
+  communityTitle: "אודות קהילת קארמה (Karma Community)",
+  whatIsTitle: "מהי קהילת קארמה?",
+  aspirationsTitle: "השאיפות המרכזיות שלנו",
+  challengesTitle: "האתגרים הקיימים כיום (הבעיות שאנו פותרים)",
+  visionTitle: "החזון והפתרונות שאנו מציעים",
+  duplicatesTitle: "כפילויות",
+  reliabilityTitle: "חוסר אמינות",
+  dispersionTitle: "פיזור",
+
+  // Categories
+  education: "חינוך",
+  health: "בריאות",
+  welfare: "רווחה",
+  environment: "סביבה",
+  animals: "בעלי חיים",
+  youthAtRisk: "נוער בסיכון",
+  elderly: "קשישים",
+  disabled: "נכים",
+  patients: "חולים",
+  familiesInNeed: "משפחות במצוקה",
+  newImmigrants: "עולים חדשים",
+  community: "קהילה",
+  work: "עבודה",
+  studies: "לימודים",
+  leisure: "בילוי",
+
+  // Cities
+  telAviv: "תל אביב",
+  jerusalem: "ירושלים",
+  haifa: "חיפה",
+  beerSheva: "באר שבע",
+  eilat: "אילת",
+  ashdod: "אשדוד",
+  ramatGan: "רמת גן",
+  petahTikva: "פתח תקווה",
+
+  // Sort Options
+  alphabetical: "אלפביתי",
+  byLocation: "לפי מיקום",
+  byCategory: "לפי תחום",
+  byPrice: "לפי מחיר",
+  byTime: "לפי שעה",
+  byRating: "לפי דירוג",
+  byRelevance: "לפי רלוונטיות",
+  byEstablishmentDate: "לפי תאריך הקמה",
+  byDonorsCount: "לפי מספר תורמים",
+
+  // Trump Specific
+  departureTime: "שעת יציאה",
+  selectDepartureTime: "בחר שעת יציאה",
+  searchAvailableTrumps: "חפש טרמפים זמינים",
+  oneWay: "בחד יעד",
+  searchCharitiesForDonation: "חפש עמותות לתרומה",
+  searchCharitiesForHelp: "חפש עמותות שיכולות לעזור לך מבחינה כספית",
+  searchCoursesAndLessons: "חפש קורסים ושיעורים",
+
+  // WhatsApp Groups
+  trumpsCenter: "טרמפים מרכז",
+  trumpsNorth: "טרמפים צפון",
+  trumpsSouth: "טרמפים דרום",
+
+  // Group Types
+  whatsapp: "whatsapp",
+  facebook: "facebook",
+
+  // Driver Names (for demo data)
+  davidCohen: "דוד כהן",
+  saraLevy: "שרה לוי",
+  mosheGoldberg: "משה גולדברג",
+  rachelAbraham: "רחל אברהם",
+  yossiShapira: "יוסי שפירא",
+  michalRosen: "מיכל רוזן",
+
+  // Organization Names (for demo data)
+  goldenHeart: "לב זהב - עמותה לתמיכה בקשישים",
+  lightForChildren: "אור לילדים - עמותה לקידום חינוך",
+  compassionForAnimals: "חמלה לבעלי חיים - עמותה להצלת חיות",
+  healthForAll: "בריאות לכולם - קידום רפואה נגישה",
+  greenEyes: "ירוק בעיניים - שמירה על איכות הסביבה",
+  bellsOfHope: "פעמוני תקווה - תמיכה בנוער בסיכון",
+  pathOfLight: "שביל האור - ליווי אנשים עם מוגבלויות",
+  voiceOfHope: "קול התקווה - תמיכה בחולי סרטן",
+  warmHome: 'בית אבות "בית חם"',
+  goldenHeartShelter: 'מקלט חיות "לב זהב"',
+  gatesOfRighteousness: 'בית חולים "שערי צדק"',
+
+  // Descriptions (for demo data)
+  elderlySupportDesc: "תמיכה בקשישים בודדים וסיוע יומיומי",
+  educationPromotionDesc: "קידום חינוך לילדים ממשפחות מעוטות יכולת",
+  animalRescueDesc: "הצלה וטיפול בבעלי חיים נטושים",
+  accessibleMedicineDesc: "קידום רפואה נגישה לכל האוכלוסיות",
+  environmentalProtectionDesc: "שמירה על איכות הסביבה והטבע",
+  youthAtRiskSupportDesc: "תמיכה וטיפול בנוער בסיכון",
+  disabledSupportDesc: "ליווי ושילוב אנשים עם מוגבלויות",
+  cancerSupportDesc: "תמיכה נפשית ופיזית בחולי סרטן",
+
+  // Search Results
+  noResultsFound: "לא נמצאו תוצאות",
+  tryChangingSearchTerms: "נסה לשנות את מילות החיפוש או לבדוק את האיות",
+  searching: "מחפש...",
+  resultsFound: "נמצאו {count} תוצאות עבור \"{query}\"",
+  noOptions: "אין תוצאות",
+
+  // Section Titles
+  realTimeData: "נתונים בזמן אמת",
+  communityImpact: "השפעה קהילתית",
+  topContributors: "תורמים מובילים החודש",
+  weeklyActivity: "פעילות שבועית",
+  quickActions: "פעולות מהירות",
+  activityGraphPlaceholder: "גרף פעילות יופיע כאן",
+
+  // Impact Stats
+  sharedRides: "נסיעות שותפו החודש",
+  donatedMeals: "ארוחות נתרמו",
+  mentoringHours: "שעות חונכות",
+  supportedFamilies: "משפחות נתמכו",
+
+  // Bubble Component Text
+  bubbleTitle: "הקהילה במספרים",
+
+  // Settings Screen
+  settingsTitle: "הגדרות",
+  settingsSubtitle: "הגדרות האפליקציה",
+  
+  // Settings Sections
+  accountSection: "חשבון",
+  notificationsSection: "התראות", 
+  appPreferencesSection: "העדפות אפליקציה",
+  helpSupportSection: "עזרה ותמיכה",
+  
+  // Settings Items
+  editProfile: "עריכת פרופיל",
+  privacy: "פרטיות",
+  privacyVisibility: "הצג סטטוס מקוון",
+  pushNotifications: "התראות דחיפה",
+  emailNotifications: "התראות דוא\"ל",
+  darkMode: "מצב כהה",
+  dataSaver: "חיסכון בנתונים",
+  clearCache: "ניקוי מטמון",
+  helpCenter: "מרכז העזרה",
+  about: "אודות",
+  logout: "התנתק",
+  
+  // Settings Alerts
+  clearCacheTitle: "ניקוי מטמון",
+  clearCacheMessage: "האם אתה בטוח שברצונך לנקות את המטמון של האפליקציה? פעולה זו תפנה שטח אחסון.",
+  cacheClearedTitle: "המטמון נוקה",
+  cacheClearedMessage: "המטמון של האפליקציה נוקה בהצלחה.",
+  logoutTitle: "התנתק",
+  logoutMessage: "האם אתה בטוח שברצונך להתנתק?",
+  linkErrorTitle: "שגיאה",
+  linkErrorMessage: "לא ניתן לפתוח את הקישור",
+  
+  // Common Actions
+  cancel: "ביטול",
+  clear: "נקה",
+
+  // === טקסטים מ-HomeScreen.tsx ===
+  // Bubble labels
+  moneyDonations: "תרומות כספיות",
+  foodKg: 'ק"ג מזון',
+  clothingKg: 'ק"ג בגדים',
+  bloodLiters: "ליטר דם",
+  volunteerHours: "שעות התנדבות",
+  rides: "טרמפים",
+  courses: "קורסים",
+  treesPlanted: "עצים ניטעו",
+  animalsAdopted: "חיות אומצו",
+  events: "אירועים",
+  recyclingBags: "שקיות מיחזור",
+  culturalEvents: "אירועי תרבות",
+  doctorVisits: "ביקורי רופא",
+  elderlySupportCount: "קשישים נתמכים",
+  childrenSupportCount: "ילדים נתמכים",
+  sportsGroups: "קבוצות ספורט",
+  musicLessons: "שיעורי מוזיקה",
+  artWorkshops: "סדנאות יצירה",
+  computerLessons: "שיעורי מחשב",
+  communityGardens: "גינות קהילתיות",
+  communityLeaderships: "הנהגות קהילתיות",
+
+  // === טקסטים מ-ProfileScreen.tsx ===
+  // Alert titles and descriptions
+  post: "פוסט",
+  postNumber: "פוסט מספר {{number}}",
+  newUser: "משתמש חדש",
+  selectedUser: "נבחר המשתמש: {{name}}",
+  shareProfile: "שתף פרופיל",
+  shareProfileDesc: "שיתוף הפרופיל שלך",
+  editProfile: "ערוך פרופיל", 
+  editProfileDesc: "עריכת פרטי הפרופיל",
+  settings: "הגדרות",
+  openSettings: "פתיחת הגדרות",
+  help: "עזרה",
+  openHelp: "פתיחת עזרה",
+  activity: "פעילות",
+  viewActivity: "צפייה בפעילות שלך",
+  history: "היסטוריה",
+  activityHistory: "היסטוריית פעילות",
+  favorites: "מועדפים",
+  yourFavorites: "המועדפים שלך",
+  discoverPeople: "גילוי אנשים",
+  findNewPeople: "מציאת אנשים חדשים",
+  highlight: "היילייט",
+  highlightNumber: "היילייט {{number}}",
+
+  // === טקסטים מ-MoneyScreen.tsx ===
+  searchCharitiesForDonation: "חפש עמותות לתרומה",
+  searchCharitiesForHelp: "חפש עמותות לעזרה",
+
+  // === טקסטים מ-KnowledgeScreen.tsx ===
+  searchCoursesAndLessons: "חפש קורסים ושיעורים",
+  knowledgeDonationTitle: "תרומת ידע",
 };
 
 // פונקציה להחלפת משתנים בטקסט
